whatpm/t/tree-test-flow.dat

#data
<p><article></article>
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <p>
|     <article>

#data
<p><aside></aside>
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <p>
|     <aside>

#data
<p><datagrid></datagrid>
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <p>
|     <datagrid>

#data
<p><dialog></dialog>
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <p>
|     <dialog>

#data
<p><footer></footer>
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <p>
|     <footer>

#data
<p><header></header>
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <p>
|     <header>

#data
<p><nav></nav>
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <p>
|     <nav>

#data
<p><section></section>
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <p>
|     <section>

#data
<article><span></article>x
#errors
no DOCTYPE
no </span>
#document
| <html>
|   <head>
|   <body>
|     <article>
|       <span>
|     "x"

#data
<article><p></article>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <article>
|       <p>
|     "x"

#data
<article><marquee></article>x
#errors
no DOCTYPE
unexpected end tag
no </marquee>
#document
| <html>
|   <head>
|   <body>
|     <article>
|       <marquee>
|         "x"

#data
<aside><span></aside>x
#errors
no DOCTYPE
no </span>
#document
| <html>
|   <head>
|   <body>
|     <aside>
|       <span>
|     "x"

#data
<aside><p></aside>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <aside>
|       <p>
|     "x"

#data
<aside><marquee></aside>x
#errors
no DOCTYPE
unexpected end tag
no </marquee>
#document
| <html>
|   <head>
|   <body>
|     <aside>
|       <marquee>
|         "x"

#data
<datagrid><span></datagrid>x
#errors
no DOCTYPE
no </span>
#document
| <html>
|   <head>
|   <body>
|     <datagrid>
|       <span>
|     "x"

#data
<datagrid><p></datagrid>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <datagrid>
|       <p>
|     "x"

#data
<datagrid><marquee></datagrid>x
#errors
no DOCTYPE
unexpected end tag
no </marquee>
#document
| <html>
|   <head>
|   <body>
|     <datagrid>
|       <marquee>
|         "x"

#data
<details><span></details>x
#errors
no DOCTYPE
no </span>
#document
| <html>
|   <head>
|   <body>
|     <details>
|       <span>
|     "x"

#data
<details><p></details>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <details>
|       <p>
|     "x"

#data
<details><marquee></details>x
#errors
no DOCTYPE
unexpected end tag
no </marquee>
#document
| <html>
|   <head>
|   <body>
|     <details>
|       <marquee>
|         "x"

#data
<dialog><span></dialog>x
#errors
no DOCTYPE
no </span>
#document
| <html>
|   <head>
|   <body>
|     <dialog>
|       <span>
|     "x"

#data
<dialog><p></dialog>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <dialog>
|       <p>
|     "x"

#data
<dialog><marquee></dialog>x
#errors
no DOCTYPE
unexpected end tag
no </marquee>
#document
| <html>
|   <head>
|   <body>
|     <dialog>
|       <marquee>
|         "x"

#data
<figure><span></figure>x
#errors
no DOCTYPE
no </span>
#document
| <html>
|   <head>
|   <body>
|     <figure>
|       <span>
|     "x"

#data
<figure><p></figure>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <figure>
|       <p>
|     "x"

#data
<figure><marquee></figure>x
#errors
no DOCTYPE
unexpected end tag
no </marquee>
#document
| <html>
|   <head>
|   <body>
|     <figure>
|       <marquee>
|         "x"

#data
<footer><span></footer>x
#errors
no DOCTYPE
no </span>
#document
| <html>
|   <head>
|   <body>
|     <footer>
|       <span>
|     "x"

#data
<footer><p></footer>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <footer>
|       <p>
|     "x"

#data
<footer><marquee></footer>x
#errors
no DOCTYPE
unexpected end tag
no </marquee>
#document
| <html>
|   <head>
|   <body>
|     <footer>
|       <marquee>
|         "x"

#data
<header><span></header>x
#errors
no DOCTYPE
no </span>
#document
| <html>
|   <head>
|   <body>
|     <header>
|       <span>
|     "x"

#data
<header><p></header>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <header>
|       <p>
|     "x"

#data
<header><marquee></header>x
#errors
no DOCTYPE
unexpected end tag
no </marquee>
#document
| <html>
|   <head>
|   <body>
|     <header>
|       <marquee>
|         "x"

#data
<nav><span></nav>x
#errors
no DOCTYPE
no </span>
#document
| <html>
|   <head>
|   <body>
|     <nav>
|       <span>
|     "x"

#data
<nav><p></nav>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <nav>
|       <p>
|     "x"

#data
<nav><marquee></nav>x
#errors
no DOCTYPE
unexpected end tag
no </marquee>
#document
| <html>
|   <head>
|   <body>
|     <nav>
|       <marquee>
|         "x"
